fix: move testing-helpers to tests/helpers/mock-pipeline
