+++ nodejs/src/ingestion/pipelines/base-batch-pipeline.test.ts
@@ -1,9 +1,9 @@
 import { Message } from 'node-rdkafka'
 
+import { createMockPipeline } from '../../../tests/helpers/mock-pipeline'
 import { BaseBatchPipeline } from './base-batch-pipeline'
 import { createBatch, createContext, createNewBatchPipeline, createOkContext } from './helpers'
 import { dlq, drop, ok } from './results'
-import { createMockPipeline } from './testing-helpers'
 
 function createTestMessage(overrides: Partial<Message> = {}): Message {
     return {
