+++ nodejs/src/ingestion/pipelines/batch-pipeline-unwrapper.test.ts
@@ -1,12 +1,12 @@
 import { Message } from 'node-rdkafka'
 
+import { createMockPipeline } from '../../../tests/helpers/mock-pipeline'
 import { logger } from '../../utils/logger'
 import { OVERFLOW_OUTPUT } from '../common/outputs'
 import { BatchPipelineUnwrapper } from './batch-pipeline-unwrapper'
 import { BatchPipeline } from './batch-pipeline.interface'
 import { DefaultContext, createContext, createNewBatchPipeline, createOkContext } from './helpers'
 import { dlq, drop, ok, redirect } from './results'
-import { createMockPipeline } from './testing-helpers'
 
 // Mock the logger
 jest.mock('../../utils/logger', () => ({
