+++ nodejs/src/ingestion/pipelines/filter-map-batch-pipeline.test.ts
@@ -1,11 +1,11 @@
 import { Message } from 'node-rdkafka'
 
 import { createTestMessage } from '../../../tests/helpers/kafka-message'
+import { createMockPipeline } from '../../../tests/helpers/mock-pipeline'
 import { BatchPipelineResultWithContext } from './batch-pipeline.interface'
 import { FilterMapBatchPipeline, FilterMapMappingFunction } from './filter-map-batch-pipeline'
 import { createContext, createNewBatchPipeline, createOkContext } from './helpers'
 import { dlq, drop, ok } from './results'
-import { createMockPipeline } from './testing-helpers'
 
 describe('FilterMapBatchPipeline', () => {
     let message1: Message
