+++ nodejs/src/ingestion/pipelines/side-effect-handling-pipeline.test.ts
@@ -1,5 +1,6 @@
 import { Message } from 'node-rdkafka'
 
+import { createMockPipeline } from '../../../tests/helpers/mock-pipeline'
 import { createBatch, createContext, createNewBatchPipeline, createOkContext } from './helpers'
 import { sideEffectResultCounter } from './metrics'
 import { dlq, drop, ok } from './results'
@@ -8,7 +9,6 @@ import {
     SideEffectHandlingConfig,
     SideEffectHandlingPipeline,
 } from './side-effect-handling-pipeline'
-import { createMockPipeline } from './testing-helpers'
 
 // Mock the metrics module
 jest.mock('./metrics', () => ({
