+++ nodejs/tests/helpers/mock-pipeline.ts
@@ -1,6 +1,6 @@
 import { Message } from 'node-rdkafka'
 
-import { BatchPipeline, BatchPipelineResultWithContext } from './batch-pipeline.interface'
+import { BatchPipeline, BatchPipelineResultWithContext } from '../../src/ingestion/pipelines/batch-pipeline.interface'
 
 /**
  * Creates a mock pipeline that returns pre-built results from next().
